added handlebars for shoppingCart
added handlebars template for shoppingCart
added columns product.name, product.price, product.update_at,
product.quantity

product.update_at could be a reference for when the item was added
to the cart

commented out the update product quantity button. We can uncomment this
when we decide on adding the quantity feature but this should be done
after we accomplish the mvp.

The remove button uses the logic that I had implemented in my second
project. By placing data-id="{{product.id}} on the handlebars button,
with some event logic, the id can be grabbed from the product and
passed through the api call then deleted. This implicit use of id
allows a user to simply press the remove item button with no further
user intervention.