added update quantity to dataTable
In order to change the quantity of an item in the shopping cart a user
must click the 'update quantity' button. This button triggers a modal,
in which the user will enter the desired quantity. This button lives
in the shoppingCart.handlebars file. IT is currently commented out, as
this is a stretch goal.
The body of this modal lives in index.html and is also commented out.

If or when we get to this feature, uncomment the respective sections and
add to then focus on setting up the appropriate logic in the events, api
and ui files.