Implement Add to Cart feature
Adds a cart variable to products/ui.js that takes in values
for id and quantity as an array when a user adds an item to their
cart from the products page.

Also adds a products table to HTML and a products handlebars file
to properly display products and the add to cart form when the show all
products button is hit.

Coding partner: Ryan K