Add to cart feature working
Add to cart button successfully sends a patch request to update the
product field with the user's shopping cart.

This commit is a long time coming- It took a very long time and a lot
of troubleshooting only to discover that the problem was a plural
where a singular should have been on the front end.

Expected conflicts with Stripe integration, but we will work them
out as a team.