Fix new user cart bug
Fixed bug that would not create a new cart for a new user when
the account is created after a previous session had been signed out.

Fixed by setting cart.currentOrder to null rather than {} so that
it would be falsy in the conditional and the new cart request would
be fired.